requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
et-xmlfile==1.1.0
fastapi==0.68.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.2.0
idna==3.2
numpy==1.21.2
openai==0.10.3
openpyxl==3.0.7
pandas==1.3.2
pandas-stubs==1.2.0.13
promise==2.3
pydantic==1.8.2
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
Rx==1.6.1
six==1.16.0
starlette==0.14.2
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
